+++ lib/src/core/constants/constants.dart
@@ -2,5 +2,4 @@
 library;
 
 export 'api_endpoints.dart';
-export 'app_strings.dart';
 export 'error_messages.dart';