+++ lib/src/presentation/controller/controller.dart
@@ -1,4 +1,0 @@
-/// Controladores de la capa de presentación.
-library;
-
-export 'application_controller.dart';